Touching up on org work...
Almost done with this API in terms of audio analyzing.

Our BPM range is usually ~1 bpm off which is extremely accurate and something I am happy with. When testing with S&M the key value was off no matter what changes I've made, but testing it with other songs it works well so I can only assume its due to the song file being downloaded poorly.

Built our gemini.py, GeminiClient class, which builds the api with the key n model and then builds a response.

Next step is to build our backend using flaskAPI and then the frontend.